+++ autorest/codegen/models/parameter.py
@@ -55,6 +55,7 @@ class Parameter(BaseModel):  # pylint: disable=too-many-instance-attributes
         constraints: List[Any],
         target_property_name: Optional[Union[int, str]] = None,  # first uses id as placeholder
         style: Optional[ParameterStyle] = None,
+        explode: Optional[bool] = False,
         *,
         flattened: bool = False,
         grouped_by: Optional["Parameter"] = None,
@@ -73,6 +74,7 @@ class Parameter(BaseModel):  # pylint: disable=too-many-instance-attributes
         self.constraints = constraints
         self.target_property_name = target_property_name
         self.style = style
+        self.explode = explode
         self.flattened = flattened
         self.grouped_by = grouped_by
         self.original_parameter = original_parameter
@@ -195,6 +197,7 @@ class Parameter(BaseModel):  # pylint: disable=too-many-instance-attributes
             constraints=[],  # FIXME constraints
             target_property_name=id(yaml_data["targetProperty"]) if yaml_data.get("targetProperty") else None,
             style=ParameterStyle(http_protocol["style"]) if "style" in http_protocol else None,
+            explode=http_protocol.get("explode", False),
             grouped_by=yaml_data.get("groupedBy", None),
             original_parameter=yaml_data.get("originalParameter", None),
             flattened=yaml_data.get("flattened", False),