+++ test/vanilla/Expected/AcceptanceTests/BodyNumber/bodynumber/aio/operations_async/_number_operations_async.py
@@ -220,7 +220,7 @@ class NumberOperations:
     ) -> None:
         """Put big float value 3.402823e+20.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -310,7 +310,7 @@ class NumberOperations:
     ) -> None:
         """Put big double value 2.5976931e+101.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -576,7 +576,7 @@ class NumberOperations:
     ) -> None:
         """Put big decimal value 2.5976931e+101.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -842,7 +842,7 @@ class NumberOperations:
     ) -> None:
         """Put small float value 3.402823e-20.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -932,7 +932,7 @@ class NumberOperations:
     ) -> None:
         """Put small double value 2.5976931e-101.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -1022,7 +1022,7 @@ class NumberOperations:
     ) -> None:
         """Put small decimal value 2.5976931e-101.
 
-        :param number_body:
+        :param number_body: number body.
         :type number_body: float
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)