+++ test/vanilla/Expected/AcceptanceTests/UrlMultiCollectionFormat/urlmulticollectionformat/operations/_queries_operations.py
@@ -70,7 +70,7 @@ class QueriesOperations(object):
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
         if array_query is not None:
-            query_parameters['arrayQuery'] = self._serialize.query("array_query", array_query, '[str]', div=',')
+            query_parameters['arrayQuery'] = [self._serialize.query("array_query", q, 'str') if q is not None else '' for q in array_query]
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
@@ -98,6 +98,8 @@ class QueriesOperations(object):
         # type: (...) -> None
         """Get an empty array [] of string using the multi-array format.
 
+        No query parameter should be sent, since the array is empty.
+
         :param array_query: an empty array [] of string using the multi-array format.
         :type array_query: list[str]
         :keyword callable cls: A custom type or function that will be passed the direct response
@@ -115,7 +117,7 @@ class QueriesOperations(object):
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
         if array_query is not None:
-            query_parameters['arrayQuery'] = self._serialize.query("array_query", array_query, '[str]', div=',')
+            query_parameters['arrayQuery'] = [self._serialize.query("array_query", q, 'str') if q is not None else '' for q in array_query]
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
@@ -141,11 +143,13 @@ class QueriesOperations(object):
         **kwargs  # type: Any
     ):
         # type: (...) -> None
-        """Get an array of string ['ArrayQuery1', 'begin!*'();:@ &=+$,/?#[]end' , null, ''] using the
-        mult-array format.
+        """Get an array of string using the multi-array format.
+
+        Parameter is ['ArrayQuery1', 'begin!*'();:@ &=+$,/?#[]end' , null, '']. null is sent as empty
+        string.
 
         :param array_query: an array of string ['ArrayQuery1', 'begin!*'();:@ &=+$,/?#[]end' , null,
-         ''] using the mult-array format.
+         ''] using the multi-array format.
         :type array_query: list[str]
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None, or the result of cls(response)
@@ -162,7 +166,7 @@ class QueriesOperations(object):
         # Construct parameters
         query_parameters = {}  # type: Dict[str, Any]
         if array_query is not None:
-            query_parameters['arrayQuery'] = self._serialize.query("array_query", array_query, '[str]', div=',')
+            query_parameters['arrayQuery'] = [self._serialize.query("array_query", q, 'str') if q is not None else '' for q in array_query]
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]